Modify the testing framework.
