+++ test/cppunit/test_main.cpp
@@ -1,6 +1,6 @@
 
-#include "../allocator_test.h"
 #include "cppunit_mini.h"
+#include "cppunit_until.h"
 #include "file_reporter.h"
 #include <stdio.h>
